backend/requirements.txt
Flask==3.1.0
Flask-Limiter==3.10.1
flask-cors==5.0.0
gunicorn==23.0.0
joblib==1.4.2
nltk==3.9.1
numpy==2.2.2
pandas==2.2.3
python-dotenv==1.1.0
pytest==8.4.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
